# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: GradeEscolar — Sistema Completo de Gestão de Horários Escolares
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /turmas:
    get:
      operationId: listTurmas
      tags: [turmas]
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Turma"
    post:
      operationId: createTurma
      tags: [turmas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TurmaInput"
      responses:
        "201":
          description: Criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turma"
  /turmas/{id}:
    get:
      operationId: getTurma
      tags: [turmas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Turma
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turma"
        "404":
          description: Não encontrada
    patch:
      operationId: updateTurma
      tags: [turmas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TurmaUpdate"
      responses:
        "200":
          description: Atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turma"
    delete:
      operationId: deleteTurma
      tags: [turmas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removida
  /turmas/{id}/aplicar-matriz:
    post:
      operationId: aplicarMatrizTurma
      tags: [turmas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AplicarMatrizInput"
      responses:
        "200":
          description: Turma com disciplinas e cargas horárias substituídas pelas da matriz (RF-TUR-02)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turma"
        "404":
          description: Turma ou matriz curricular não encontrada
  /turmas/{id}/horario:
    get:
      operationId: getTurmaHorario
      tags: [turmas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grade horária
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HorarioSlot"

  # ── HORÁRIOS ──────────────────────────────────────────────────────────────────
...
  /usuarios:
    get:
      operationId: listUsuarios
      tags: [usuarios]
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
    post:
      operationId: createUsuario
      tags: [usuarios]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioInput"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
  /usuarios/me:
    get:
      operationId: getMe
      tags: [usuarios]
      summary: Retorna o perfil do usuário autenticado
      responses:
        "200":
          description: Meu perfil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "404":
          description: Não cadastrado ainda
  /usuarios/{id}:
    patch:
      operationId: updateUsuario
      tags: [usuarios]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioUpdate"
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"

  # ── AUDIT LOG ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Professor:
      type: object
      required: [id, nome, email, ativo, createdAt]
      properties:
        id: { type: integer }
        nome: { type: string }
        email: { type: string }
        telefone: { type: ["string", "null"] }
        ativo: { type: boolean }
        disciplinaIds:
          type: array
          items: { type: integer }
        createdAt: { type: string }
...
    Disciplina:
      type: object
      required: [id, nome, cargaSemanal, cor, createdAt]
      properties:
        id: { type: integer }
        nome: { type: string }
        cargaSemanal: { type: integer }
        cor: { type: string }
        createdAt: { type: string }
...
    Turma:
      type: object
      required: [id, nome, serie, turno, anoLetivo, createdAt]
      properties:
        id: { type: integer }
        nome: { type: string }
        serie: { type: string }
        turno: { type: string, enum: [matutino, vespertino, noturno, integral] }
        anoLetivo: { type: integer }
        modalidade: { type: ["string", "null"] }
        matrizCurricularId: { type: ["integer", "null"], description: "Matriz Curricular aplicada a esta turma (RF-TUR-02), se houver." }
        disciplinaIds:
          type: array
          items: { type: integer }
        disciplinasComCarga:
          type: array
          description: "Carga horária semanal efetiva por disciplina (override da matriz aplicada, com fallback para a carga global)."
          items:
            type: object
            required: [disciplinaId, nome, cargaHorariaSemanal, origemMatriz]
            properties:
              disciplinaId: { type: integer }
              nome: { type: string }
              cargaHorariaSemanal: { type: integer }
              origemMatriz: { type: boolean }
        createdAt: { type: string }
    AplicarMatrizInput:
      type: object
      required: [matrizCurricularId]
      properties:
        matrizCurricularId: { type: integer }
    TurmaInput:
      type: object
      required: [nome, serie, turno, anoLetivo]
      properties:
        nome: { type: string, minLength: 1 }
        serie: { type: string }
        turno: { type: string, enum: [matutino, vespertino, noturno, integral] }
        anoLetivo: { type: integer }
        modalidade: { type: string }
        disciplinaIds:
          type: array
          items: { type: integer }
    TurmaUpdate:
      type: object
      properties:
        nome: { type: string }
        serie: { type: string }
        turno: { type: string, enum: [matutino, vespertino, noturno, integral] }
        anoLetivo: { type: integer }
        modalidade: { type: string }
        disciplinaIds:
          type: array
          items: { type: integer }
    HorarioSlot:
      type: object
      required: [id, turmaId, disciplinaId, professorId, diaSemana, numeroAula]
      properties:
        id: { type: integer }
        turmaId: { type: integer }
        turma:
          $ref: "#/components/schemas/Turma"
        disciplinaId: { type: integer }
        disciplina:
          $ref: "#/components/schemas/Disciplina"
        professorId: { type: integer }
        professor:
          $ref: "#/components/schemas/Professor"
        diaSemana: { type: integer, minimum: 0, maximum: 4 }
        numeroAula: { type: integer, minimum: 1, maximum: 8 }
        sala: { type: ["string", "null"] }
...
    Usuario:
      type: object
      required: [id, clerkId, nome, email, perfil, ativo]
      properties:
        id: { type: integer }
        clerkId: { type: string }
        nome: { type: string }
        email: { type: string }
        perfil: { type: string, enum: [direcao, coordenacao, professor, secretaria] }
        ativo: { type: boolean }
        createdAt: { type: string }
    UsuarioInput:
      type: object
      required: [clerkId, nome, email]
      properties:
        clerkId: { type: string }
        nome: { type: string }
        email: { type: string }
        perfil: { type: string, enum: [direcao, coordenacao, professor, secretaria] }
    UsuarioUpdate:
      type: object
      properties:
        nome: { type: string }
        perfil: { type: string, enum: [direcao, coordenacao, professor, secretaria] }
        ativo: { type: boolean }
...